\name{acceptance}
\alias{acceptance}
\title{List the history of acceptance.}
\description{
 This function reads "accept.xml" in .Library, if it exists, and prints it
 with the default method.
}
\usage{
acceptance()
}
\details{
  Data is displayed as xml markup.
}
\value{
 an object of class XMLNode
  ...
}
\references{http://metruminstitute.org}
\author{Tim Bergsma}
\seealso{
	\itemize{
		\item{}{\code{\link{accept}}}
		\item{}{\code{\link{packageCheck}}}
	}
}
\keyword{ manip}

