\name{as.miDate.miDate}
\alias{as.miDate.miDate}
\title{Create an miDate object from an miDate object}
\description{Create an miDate object from an miDate object.}
\usage{
\method{as.miDate}{miDate}(x, ...)
}
\arguments{
  \item{x}{miDate}
  \item{\dots}{ignored arguments}
}
\details{
  Consistency utility that returns its argument.
}
\value{
miDate
}
\references{http://metruminstitute.org}
\author{Tim Bergsma}
\seealso{
\code{\link{as.miDate}}}
\examples{
function(x,...)return(x)
}
\keyword{manip}
