\name{check.data}
\alias{check.data}
\title{Summarize columns in a data set}
\description{
  This function checks for missing data, numeric data, ranges, etc.
  It is a tool developed to review data that has been read into R.
}
\usage{
check.data(IDname, data)
}
\arguments{
  \item{IDname}{subject identifier in the data.}
  \item{data}{name of R object}
}
\details{
 }
\value{
  A summary of each column in the \code{data} by \code{IDname} is output.
  }
\references{ MIfuns package available from Metrum Institute website (\code{http://metruminstitute.org/}). }
\author{ Developed by Metrum Institute (Bill Knebel, Tim Bergsma, and others). }
\seealso{ }
\examples{
}
\keyword{ manip }
