\name{CheckPositiveVector}
\alias{CheckPositiveVector}
\alias{CheckPositiveMatrix}
\title{Verify that Omega matrix is positive, definite}
\description{
  Helper function for \code{SimulationFromFile} that determines if
  omega matrix is positive, definite. 
  }
\usage{
CheckPositiveVector(omegaVector)
}

\arguments{
  \item{omegaVector}{R vector containing omega values}
}
\details{
}
\value{
}
\references{ MIfuns package available from Metrum Institute website (\code{http://metruminstitute.org/}). }
\author{ Developed by Metrum Institute (Bill Knebel, Tim Bergsma, and others). }
\seealso{ }
\examples{
}
\keyword{ manip }

