\name{CheckWS}
\alias{CheckWS}
\title{Check file name for white space}
\description{
 Helper function to check if a file name does contains white space.
}
\usage{
CheckWS(x)
}
\arguments{
  \item{x}{checks if \code{x} contains white space}
}
\value{
  returns TRUE if \code{x} contains white space and FALSE if it does not.
}
\references{ MIfuns package available from Metrum Institute website (\code{http://metruminstitute.org/}). }
\author{ Developed by Metrum Institute (Bill Knebel, Tim Bergsma, and others). }
\seealso{ }
\examples{
}
\keyword{ manip }
