\name{getPars}
\alias{getPars}

\title{Return the Non-duplicated (ID) Rows Corresponding to the Specified Filename}
\description{
 Useful for retrieving parameters that are constant within subject.
}
\usage{
getPars(file)
}
\arguments{
  \item{file}{path for a file, typ. *par.TAB}
}
\value{data.frame}
\references{mifuns.googlecode.com}
\author{Tim Bergsma}
\seealso{ \code{\link{PLOTR}} }
\keyword{manip}

