\name{helpAdminister}
\alias{helpAdminister}
\docType{package}
\title{MIfuns: Administrative}
\description{Tools for administrative tasks.}

\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item \link[=CLNR]{delete} files and directories; log what was deleted to a text file
\item \link[=hash]{embed} R output in a script
\item \link[=accept]{document} acceptance of an R installation
\item \link[=variants]{locate} path-variant versions of a file
 }
}

