\name{helpClasses}
\alias{helpClasses}
\docType{package}
\title{\link[=MIfuns]{MIfuns} \link[=helpPrepare]{Data}: Data Classes}
\description{Classes for building feature-rich data.}
\details{
I want to ....
\itemize{
\item \link[=as.comment]{Create} and manipulate comment objects
\item \link[=as.flag]{Create} and manipulate flags
\item \link[=as.keyed]{Create} and manipulate keyed data.frames
\item \link[=Ops.keyed]{Join} keyed data.frames
\item \link[=as.nm]{Create} and manipulate nm objects
\item \link[=as.moot]{Modify} the assembly of nm objects
}
}
