\name{helpList}
\alias{helpList}
\docType{package}
\title{
Functions defined by MIfuns
}
\details{
\itemize{
\item \code{\link{\%nested.in\%}}
\item \code{\link{\%nests\%}}
\item \code{\link{AUC}}
\item \code{\link{CLNR}}
\item \code{\link{CheckPositiveMatrix}}
\item \code{\link{CheckPositiveVector}}
\item \code{\link{ComputePredictiveCheck}}
\item \code{\link{CreateParametersForSimulation}}
\item \code{\link{NONR}}
\item \code{\link{Omega.MtoV}}
\item \code{\link{Omega.VtoM}}
\item \code{\link{Ops.keyed}}
\item \code{\link{PLOTR}}
\item \code{\link{PlotPredictiveCheck}}
\item \code{\link{SimulateOmega}}
\item \code{\link{TABR}}
\item \code{\link{Tmax}}
\item \code{\link{Tmin}}
\item \code{\link{TruncateParametersForSimulation}}
\item \code{\link{[.comment}}
\item \code{\link{[.flag}}
\item \code{\link{[.keyed}}
\item \code{\link{[.miTemporal}}
\item \code{\link{[<-.miTemporal}}
\item \code{\link{[[.comment}}
\item \code{\link{[[.flag}}
\item \code{\link{[[.miTemporal}}
\item \code{\link{accept}}
\item \code{\link{acceptance}}
\item \code{\link{acr}}
\item \code{\link{aggregate.keyed}}
\item \code{\link{and.keyed}}
\item \code{\link{as.character.comment}}
\item \code{\link{as.character.flag}}
\item \code{\link{as.character.miTemporal}}
\item \code{\link{as.chartime}}
\item \code{\link{as.chartime.numeric}}
\item \code{\link{as.comment}}
\item \code{\link{as.comment.comment}}
\item \code{\link{as.comment.default}}
\item \code{\link{as.csv.filename}}
\item \code{\link{as.csv.filename.character}}
\item \code{\link{as.data.frame.comment}}
\item \code{\link{as.data.frame.flag}}
\item \code{\link{as.filename}}
\item \code{\link{as.filename.character}}
\item \code{\link{as.flag}}
\item \code{\link{as.flag.default}}
\item \code{\link{as.flag.flag}}
\item \code{\link{as.keyed}}
\item \code{\link{as.keyed.data.frame}}
\item \code{\link{as.miDate}}
\item \code{\link{as.miDate.character}}
\item \code{\link{as.miDate.miDate}}
\item \code{\link{as.miDate.numeric}}
\item \code{\link{as.miDateTime}}
\item \code{\link{as.miDateTime.character}}
\item \code{\link{as.miDateTime.miDate}}
\item \code{\link{as.miDateTime.miDateTime}}
\item \code{\link{as.miDateTime.numeric}}
\item \code{\link{as.miTime}}
\item \code{\link{as.miTime.character}}
\item \code{\link{as.miTime.miTime}}
\item \code{\link{as.miTime.numeric}}
\item \code{\link{as.moot}}
\item \code{\link{as.nm}}
\item \code{\link{as.nm.data.frame}}
\item \code{\link{as.numeric.chartime}}
\item \code{\link{as.rigged}}
\item \code{\link{auc}}
\item \code{\link{aug}}
\item \code{\link{backtrans}}
\item \code{\link{badAmt}}
\item \code{\link{badAmt.nm}}
\item \code{\link{badDv}}
\item \code{\link{badDv.nm}}
\item \code{\link{badII}}
\item \code{\link{badII.nm}}
\item \code{\link{bin}}
\item \code{\link{bmi}}
\item \code{\link{bsa}}
\item \code{\link{c.comment}}
\item \code{\link{c.flag}}
\item \code{\link{c.miTemporal}}
\item \code{\link{cavg}}
\item \code{\link{check.data}}
\item \code{\link{cmax}}
\item \code{\link{cmin}}
\item \code{\link{compute.cwres}}
\item \code{\link{constant}}
\item \code{\link{constant.default}}
\item \code{\link{contains}}
\item \code{\link{covariatePlots}}
\item \code{\link{crcl}}
\item \code{\link{crosses}}
\item \code{\link{css}}
\item \code{\link{cwresPlots}}
\item \code{\link{dataFormat}}
\item \code{\link{dataSynthesis}}
\item \code{\link{deranged}}
\item \code{\link{deranged.data.frame}}
\item \code{\link{deranged.keyed}}
\item \code{\link{diagnosticPlots}}
\item \code{\link{distance}}
\item \code{\link{distance.logical}}
\item \code{\link{dupKeys}}
\item \code{\link{episcript}}
\item \code{\link{f}}
\item \code{\link{falseAmt}}
\item \code{\link{falseAmt.nm}}
\item \code{\link{falseDv}}
\item \code{\link{falseDv.nm}}
\item \code{\link{filename}}
\item \code{\link{first}}
\item \code{\link{first.logical}}
\item \code{\link{forbak}}
\item \code{\link{format.comment}}
\item \code{\link{format.flag}}
\item \code{\link{format.miDate}}
\item \code{\link{format.miDateTime}}
\item \code{\link{format.miTime}}
\item \code{\link{getCovs}}
\item \code{\link{getCwres}}
\item \code{\link{getPars}}
\item \code{\link{getTabs}}
\item \code{\link{getdname}}
\item \code{\link{grdCommand}}
\item \code{\link{groupnames}}
\item \code{\link{hash}}
\item \code{\link{hide}}
\item \code{\link{hide.data.frame}}
\item \code{\link{ibw}}
\item \code{\link{ind.cwres}}
\item \code{\link{is.cwres.readable.file}}
\item \code{\link{ke}}
\item \code{\link{key}}
\item \code{\link{key<-}}
\item \code{\link{lbm}}
\item \code{\link{left.keyed}}
\item \code{\link{locf}}
\item \code{\link{maxChar}}
\item \code{\link{merge.keyed}}
\item \code{\link{merge.nm}}
\item \code{\link{metaSub}}
\item \code{\link{metaSub.character}}
\item \code{\link{metaSub.filename}}
\item \code{\link{minus.keyed}}
\item \code{\link{minus.moot}}
\item \code{\link{moot}}
\item \code{\link{moot.nm}}
\item \code{\link{myriwish}}
\item \code{\link{naInContext}}
\item \code{\link{naKeys}}
\item \code{\link{nix}}
\item \code{\link{nm}}
\item \code{\link{nmPlots}}
\item \code{\link{noPk}}
\item \code{\link{noPk.nm}}
\item \code{\link{nocb}}
\item \code{\link{ordinal}}
\item \code{\link{ordinal.logical}}
\item \code{\link{packageCheck}}
\item \code{\link{panel.covplot}}
\item \code{\link{panel.cuts}}
\item \code{\link{panel.densitystrip}}
\item \code{\link{panel.ref}}
\item \code{\link{plot.nm}}
\item \code{\link{plus.keyed}}
\item \code{\link{plus.rigged}}
\item \code{\link{predoseDv}}
\item \code{\link{predoseDv.nm}}
\item \code{\link{prev}}
\item \code{\link{print.comment}}
\item \code{\link{print.flag}}
\item \code{\link{print.keyed.summary}}
\item \code{\link{print.miTemporal}}
\item \code{\link{print.nm.summary}}
\item \code{\link{purge.dir}}
\item \code{\link{purge.files}}
\item \code{\link{read.cwres.data}}
\item \code{\link{read.nm}}
\item \code{\link{reapply}}
\item \code{\link{regCommand}}
\item \code{\link{rep.comment}}
\item \code{\link{rep.flag}}
\item \code{\link{rep.miTemporal}}
\item \code{\link{resample}}
\item \code{\link{resample.csv.filename}}
\item \code{\link{resample.data.frame}}
\item \code{\link{resample.filename}}
\item \code{\link{rig}}
\item \code{\link{rig.nm}}
\item \code{\link{rinvchisq}}
\item \code{\link{rinvgamma}}
\item \code{\link{riwish}}
\item \code{\link{rlog}}
\item \code{\link{runCommand}}
\item \code{\link{runNonmem}}
\item \code{\link{runhead}}
\item \code{\link{safe.call}}
\item \code{\link{seq.miTemporal}}
\item \code{\link{setCwres}}
\item \code{\link{sort.keyed}}
\item \code{\link{sorted}}
\item \code{\link{sqrtm}}
\item \code{\link{stableMerge}}
\item \code{\link{star}}
\item \code{\link{strain}}
\item \code{\link{summary.keyed}}
\item \code{\link{summary.nm}}
\item \code{\link{synthesis}}
\item \code{\link{text2decimal}}
\item \code{\link{tmax}}
\item \code{\link{transform.keyed}}
\item \code{\link{unconditional}}
\item \code{\link{unconditional.data.frame}}
\item \code{\link{unconditional.default}}
\item \code{\link{uniKey}}
\item \code{\link{uniKey.keyed}}
\item \code{\link{where}}
\item \code{\link{where.logical}}
\item \code{\link{win}}
\item \code{\link{write.nm}}
\item \code{\link{xtfrm.comment}}
\item \code{\link{xtfrm.flag}}
\item \code{\link{xtfrm.miTemporal}}
\item \code{\link{zeroAmt}}
\item \code{\link{zeroAmt.nm}}
\item \code{\link{zeroDv}}
\item \code{\link{zeroDv.nm}}
}
}
  
  \author{
    Developed by Metrum Institute (Bill Knebel, Tim Bergsma, and
    others). Maintainer: Bill Knebel <billk@metrumrg.com>
  }
  
  \references{
    MIfuns package available from Metrum Institute website (\code{http://metruminstitute.org/}). }
  \keyword{
    manip }
