\name{helpMatrix}
\alias{helpMatrix}
\docType{package}
\title{\link[=MIfuns]{MIfuns} \link[=helpPrepare]{Data}: Working with Matrices}
\description{Tools for processing matrices.}
\details{
I want to ....
\itemize{
\item \link[=half.matrix]{Interconvert} between square matrices and triangular matrices 
}
}
