\name{helpPrepare}
\alias{helpPrepare}
\docType{package}
\title{
MIfuns: Data Assembly
}
\description{Tools for manipulating and assembling data for pharmacometric analysis.}
\details{
I want to ....
\itemize{
\item \link[=resample.data.frame]{Create} replicate datasets by stratified sampling with replacement    
\item \link[=metaSub.character]{Make} multiple, systematic substitutions in text files
\item \link[=miTemporal]{Create} date, time, and date-time objects with default behaviors appropriate for NMTRAN datasets
\item \link[=reapply]{Apply} a function across cells of an indexed vector, giving an identically-indexed result
\item \link[=text2decimal]{Convert} mixed text to decimal
\item \link[=forbak]{Impute} missing vector values
\item \link[=runhead]{Check} some properties of vectors and dataframes
\item \link[=bin]{Calculate} bin sizes and limits for a vector
\item \link[=bsa]{Calculate} various clinical measurements: bodysize, etc.
\item \link[=constant]{Test} for nesting among factors
\item \link[=naInContext]{View} missing values in a broader context
\item \link[=stableMerge]{Add} keyed data to a dataframe safely
\item \link[=css]{Calculate} One-Compartment Model Properties
\item \link[=as.comment]{Create} and manipulate 'comment' objects
\item \link[=as.flag]{Create} and manipulate 'flags'
\item \link[=contains]{Test} a character vector for occurences of a pattern
\item \link[=distance]{Analyze} positions in a condition; make 'vertical' inferences
\item \link[=as.keyed]{Create} and manipulate keyed data frames
\item \link[=Ops.keyed]{Join} keyed data frames
\item \link[=nm]{Create} and manipulate 'nm' objects
\item \link[=deranged]{Modify} the assembly of 'nm' objects
\item \link[=nmPlots]{Characterize} 'nm' objects with diagnostic plots
\item \link[=plot.nm]{Plot} 'nm' objects
\item \link[=summary.nm]{Summarize} an 'nm' object
\item \link[=fixedwidth]{Format} a data.frame as fixed-width
\item \link[=variants]{Locate} variants of a file in distinctive subdirectories
\item \link[=latest]{Identify} the latest of each set of variants
\item \link[=shuffle]{Move} columns in a data frame
}
}
