\name{helpStrategic}
\alias{helpStrategic}
\docType{package}
\title{MIfuns: Strategic Data Manipulation}
\description{Tools for strategic data manipulation.}
\details{
\link[=MIfuns]{MIfuns} > \link[=helpPrepare]{Data}

I want to ....
\itemize{
\item \link[=reapply]{Apply} a function across cells of an indexed vector, giving an identically-indexed result
\item \link[=stableMerge]{Add} keyed data to a data.frame safely
\item \link[=first]{Identify} elements within cells that meet some instance of a condition
}
}
