\name{Omega.MtoV}
\alias{Omega.MtoV}
\title{ Convert omega matrix to vector }
\description{
 Converts matrix containing omega estimates into a vector of omega estimates.}
\usage{
Omega.MtoV(omegaMatrix)
}
\arguments{
  \item{omegaMatrix}{ matrix containing omega estimates }
}
\details{
  This a a helper function for the \code{CheckWS} function. This is not called directly.
}
\value{
}
\references{ MIfuns package available from Metrum Institute website (\code{http://metruminstitute.org/}). }
\author{ Developed by Metrum Institute (Bill Knebel, Tim Bergsma, and others). }
\seealso{ }
\examples{
}
\keyword{utilities}
