\name{Omega.VtoM}
\alias{Omega.VtoM}
\title{Convert omega vector to matrix}
\description{
 Converts vector containing lower block omega estimates into a full
 matrix of omega estimates.}
\usage{
Omega.VtoM(omegaVector)
}
\arguments{
  \item{omegaVector}{vector of lower block omega estimates}
}
\details{
 This a a helper function for the \code{myriwish} function. This is not called directly.}
\value{
}
\references{ MIfuns package available from Metrum Institute website (\code{http://metruminstitute.org/}). }
\author{ Developed by Metrum Institute (Bill Knebel, Tim Bergsma, and others). }
\seealso{ }
\examples{
}
\keyword{utilities}
