\name{Q25}
\alias{Q25}
\title{ Function calculates 25th percentile/probability }
\description{
  Wrapper around the quantile function to calculate quantile with
  probs = 0.25
}
\usage{
Q25(x)
}
\arguments{
  \item{x}{ Vector of values passed to \code{quantile} function
  with \code{probs=c(0.25)}
  }
}
  \details{
  
}

\references{ MIfuns package available from Metrum Institute website (\code{http://metruminstitute.org/}). }
\author{ Developed by Metrum Institute (Bill Knebel, Tim Bergsma, and others). }
\examples{}
\keyword{ utilities }

