\name{Q75}
\alias{Q75}
\title{ Function calculates 75th percentile/probability }
\description{
  Wrapper around the quantile function to calculate quantile with
  probs = 0.75
}
\usage{
Q75(x)
}
\arguments{
  \item{x}{ Vector of values passed to \code{quantile} function
  with \code{probs=c(0.75)}
  }
}
  \details{
  
}

\references{ MIfuns package available from Metrum Institute website (\code{http://metruminstitute.org/}). }
\author{ Developed by Metrum Institute (Bill Knebel, Tim Bergsma, and others). }
\examples{}
\keyword{ utilities }
