\name{sqrtm}
\alias{sqrtm}
\title{ Called from compute.cwres to calculated CWRES. }
\description{
  A function called by \code{compte.cwres} to calculate CWRES. See
  help file for \code{compute.cwres}.
}
\usage{
sqrtm(x)
}
\arguments{
  \item{x}{ a matrix. }
}

\seealso{\code{\link{CWRES}} }
\examples{}
\keyword{ manip }
