\name{helpAdminister}
\alias{helpAdminister}
\docType{package}
\title{\link[=MIfuns]{MIfuns}: Administrative}
\description{Tools for administrative tasks.}

\details{
I want to ....
\itemize{
\item \link[=CLNR]{Delete} files and directories; log what was deleted to a text file
\item \link[=hash]{Embed} R output in a script
\item \link[=accept]{Document} acceptance of an R installation
\item \link[=variants]{Locate} path-variant versions of a file
 }
}

