\name{helpReport}
\alias{helpReport}
\docType{package}
\title{MIfuns: Reporting}
\description{Tools for summarizing and reporting pharmacometric analyses.}
\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item \link[=params]{identify} documented parameter names
\item \link[=lookup]{lookup} some other form of parameter names
\item \link[=rlog]{generate} a combined run log from multiple NONMEM runs
\item \link[=partab]{create} a model parameter table
}
}
