\name{helpReport}
\alias{helpReport}
\docType{package}
\title{
MIfuns: Reporting
}
\description{Tools for summarizing and reporting pharmacometric analyses.}
\details{
I want to ....
\itemize{
\item \link[=CLNR]{Delete} files and directories; log what was deleted to a text file
\item \link[=compute.cwres]{Compute} the conditional weighted residuals
\item \link[=rlog]{Generate} a combined run log from multiple NONMEM runs
\item \link[=nm]{Create} and manipulate 'nm' objects
\item \link[=nmPlots]{Characterize} 'nm' objects with diagnostic plots
\item \link[=plot.nm]{Plot} 'nm' objects
\item \link[=summary.nm]{Summarize} an 'nm' object
}
}
