\name{helpVector}
\alias{helpVector}
\docType{package}
\title{\pkg{MIfuns}: Working with Vectors}
\description{Tools for processing vectors.}.
\details{
I want to ....
\itemize{
\item \link[=text2decimal]{Convert} mixed text to decimal
\item \link[=contains]{Test} a character vector for occurences of a pattern
\item \link[=forbak]{Impute} missing vector values
\item \link[=runhead]{Check} some properties of vectors
\item \link[=miTemporal]{Create} date, time, and date-time objects
}
}
