\name{MIfuns-package}
\alias{MIfuns-package}
\alias{MIfuns}
\docType{package}
\title{
Pharmacometric tools for data preparation, analysis, simulation, and reporting
}
\description{
Pharmacometric tools for common data preparation tasks, stratified
bootstrap resampling of data sets, NONMEM control stream
creation/editing, NONMEM model execution, creation of standard and
user-defined diagnostic plots,  execution and summary of bootstrap and
predictive check results, implementation of simulations from posterior
parameter distributions, reporting of output tables and creation of a
detailed analysis log. All functions are considered stable unless marked
as beta.
}
\details{
I want to ....
\itemize{
\item \link[=helpPrepare]{prepare or manipulate data}
\item \link[=helpAnalyze]{model data in NONMEM}
\item \link[=helpSimulate]{conduct simulations}
\item \link[=helpPlot]{create diagnostic plots}
\item \link[=helpReport]{generate report material}
\item \link[=helpAdminister]{perform related filesystem tasks}
\item \link[=helpList]{see a list of available functions}

}

    
    \tabular{ll}{
Package: \tab MIfuns\cr
Type: \tab Package\cr
Version: \tab 3.9.1\cr
Date: \tab 2010-03-30\cr
License: \tab GPL\cr
}

  }
  
  \author{
    Developed by Metrum Institute (Bill Knebel, Tim Bergsma, and
    others). Maintainer: Bill Knebel <billk@metrumrg.com>
  }
  
  \references{
    MIfuns package available from Metrum Institute website (\code{http://metruminstitute.org/}). }
  \keyword{manip}
