\name{sqrtm}
\alias{sqrtm}
\title{ Called from compute.cwres to calculated CWRES. }
\description{
  A function called by \code{\link{compute.cwres}} to calculate CWRES. 
}
\usage{
sqrtm(x)
}
\arguments{
  \item{x}{ a matrix. }
}
\seealso{
	\itemize{
		\item{}{\code{\link{compute.cwres}}}
	}
}
\keyword{ manip }
